+++ src/MsSql/MsSql.ts
@@ -1,6 +1,7 @@
 import { DbCreator, type PasswordValidityTuple } from "../DbCreator";
 
 import { MsSqlPwdValidityEnum, msSqlPwdValidityEnumMessage } from "./MsSqlPwdValidityEnum";
+import { SqlInContainerRunner } from "./SqlInContainerRunner";
 import { escapeId, escapeUser, escapeStr } from "./escape";
 import { waitFor } from "./waitFor";
 
@@ -55,41 +56,26 @@ export const MsSql = new DbCreator({
 -p ${this.port}:${opts.port}\
 -d mcr.microsoft.com/mssql/server:${opts.tag}`;
 		const contId = shellOutput?.text().trim();
-
-		// TODO: parse the resposnse os SQL server and check for potential errors
-		const sqlcmd = async (sql: string, db?: string) => {
-			if (db) {
-				sql = `use ${escapeId(opts.database)}\n` + sql;
-			}
-			if (opts.verbose) {
-				console.log("SQL:", sql.includes("\n") ? "\n" + sql + " --> END SQL" : sql);
-			}
-			let prms = $`docker exec -it ${contId} \
-/opt/mssql-tools/bin/sqlcmd -S localhost \
--U SA -P ${opts.password} -Q ${sql} || exit 1`;
-			if (!opts.verbose && !opts.dryRun) {
-				prms = prms?.quiet();
-			}
-			return prms;
-		};
+		if (!contId) {
+			throw new Error("Unable to find id of the created container");
+		}
+		const sql = new SqlInContainerRunner(contId, $, opts);
 
 		console.log("Waiting for db to be up and running...");
-
 		// https://docs.docker.com/engine/reference/run/#healthchecks
-		await waitFor(() => sqlcmd("SELECT 1"));
+		await waitFor(() => sql.run("SELECT SERVERPROPERTY('ProductVersion')"), { preDelay: 1000 });
 
 		console.log("Creating the database and required data...");
-		// TODO: parse the resposnse os SQL server and check for potential errors
-		await sqlcmd(`CREATE DATABASE ${escapeId(opts.database)}`);
+		await sql.run(`CREATE DATABASE ${escapeId(opts.database)}`);
 
 		vlog("Creating login");
-		await sqlcmd(`CREATE LOGIN ${escapeUser(opts.user)} WITH PASSWORD = ${escapeStr(opts.password)}`);
+		await sql.run(`CREATE LOGIN ${escapeUser(opts.user)} WITH PASSWORD = ${escapeStr(opts.password)}`);
 
 		vlog("Creating user");
-		await sqlcmd(`create user ${escapeUser(opts.user)} for login ${escapeUser(opts.user)}`, opts.database);
+		await sql.runInDb(`create user ${escapeUser(opts.user)} for login ${escapeUser(opts.user)}`);
 
 		// To check available roles: Select	[name] From sysusers Where issqlrole = 1
 		vlog("Adding required permissions");
-		await sqlcmd(`exec sp_addrolemember 'db_owner', ${escapeStr(opts.user)}`, opts.database);
+		await sql.runInDb(`ALTER ROLE db_owner ADD MEMBER ${escapeUser(opts.user)}`);
 	},
 });