+++ src/MsSql/waitFor.ts
@@ -3,28 +3,27 @@ type MaybePromise<T> = T | Promise<T>;
 
 interface WaitForOpts {
 	signal?: AbortSignal;
-	watchDogMs?: number;
-	initialDelay?: number;
+	preDelay?: number;
+	minDelay?: number;
 	maxDelay?: number;
+	watchDogMs?: number;
 	rate?: number;
-	immediate?: boolean;
 }
 
-/** truncated exponential backoff polling, until doesn't throw */
+/** truncated exponential backoff polling, until returns truthy */
 export async function waitFor(
 	exp: () => MaybePromise<unknown>,
-	{ signal, watchDogMs = 60_000, initialDelay = 200, maxDelay = 5_000, rate = 1.5, immediate }: WaitForOpts = {}
+	{ signal, watchDogMs = 60_000, minDelay = 200, maxDelay = 5_000, rate = 1.5, preDelay }: WaitForOpts = {}
 ): Promise<void> {
 	if (!Number.isFinite(rate) || rate < 1.0 || rate > 10) {
 		throw new RangeError(`Rate for exp backof must be in range 1 <= RATE <= 10, got ${rate}`);
 	}
-	if (!Number.isInteger(initialDelay) || initialDelay < 1) {
-		throw new TypeError(`Initial delay value must be a positive integer, got ${initialDelay}`);
+	if (!Number.isInteger(minDelay) || minDelay < 1) {
+		throw new TypeError(`mindelay value must be a positive integer, got ${minDelay}`);
 	}
-	if (!Number.isInteger(maxDelay) || maxDelay < initialDelay) {
+	if (!Number.isInteger(maxDelay) || maxDelay < minDelay) {
 		throw new TypeError(
-			"Max delay value must be a positive integer bigger than initialDelay, " +
-				`got ${maxDelay} (initialDelay = ${initialDelay})`
+			"maxDelay value must be a positive integer bigger than mindelay, " + `got ${maxDelay} (mindelay = ${minDelay})`
 		);
 	}
 	if (!Number.isInteger(watchDogMs) || watchDogMs < maxDelay) {
@@ -34,22 +33,21 @@ export async function waitFor(
 		);
 	}
 
-	let currentDelay = initialDelay;
+	let currentDelay = minDelay;
 	let nRuns = 0;
-	if (!immediate) {
-		await pause(currentDelay, undefined, { signal });
-		currentDelay = Math.min(currentDelay * rate, maxDelay);
+	if (preDelay) {
+		await pause(preDelay, undefined, { signal });
 	}
 	await withTimeout(async () => {
 		for (;;) {
 			++nRuns;
 			if (signal?.aborted) {
-				throw signal.reason ?? new Error("waitFor aborted without a reason");
+				throw signal.reason ?? new Error("waitFor was aborted without a reason");
 			}
-			try {
-				await exp();
-				break;
-			} catch {}
+			// add ExecutionMax, and abort execution after it's passed, in case of long
+			// hanging promises
+			const result = await exp();
+			if (result) break;
 			await pause(currentDelay, undefined, { signal });
 			currentDelay = Math.min(currentDelay * rate, maxDelay);
 		}